# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "api documentation"
  version: "1"
  title: "cofnders api "
host: "virtserver.swaggerhub.com"
basePath: "/https://b2wuhx8v6d.execute-api.ap-southeast-1.amazonaws.com/dev"
schemes:
- "https"
paths:
  /users:
    get:
      description: "Gets All Users from Database"
      operationId: "usersGET"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "JWT token set in header"
        required: true
        type: "string"
      responses:
        200:
          description: "It should return All Users."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              users:
                type: "array"
                items:
                  type: "object"
                  properties: {}
            example:
              error: false
              users:
              - _id: "5c9efdde0bb9defc9c218c65"
                fullName: "Udit Jain"
                firstName: "Udit"
                lastName: "Jain"
                email: "jainudit.012@gmail.com"
                emailVerified: true
                profilePicture: "https://media.licdn.com/dms/image/C5103AQG_1EUajUCTXg/profile-displayphoto-shrink_100_100/0?e=1559174400&v=beta&t=WQDJMxQCMYLl4rJ2yYszirYi-rVnd2q1xvtD5greIyY"
                sub: "linkedin|oP4sPrYxft"
                isAdmin: true
                isFormFilled: true
                isActive: true
                createdAt: "1553923550413"
                updatedAt: "1553927113266"
                brief: "my brief"
                purpose: "my purpose"
        403:
          description: "Not authorized if user is not Admin."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Not Authorized!"
      x-swagger-router-controller: "Default"
    post:
      description: "Lets user get a new token"
      operationId: "usersPOST"
      parameters:
      - in: "body"
        name: "Body"
        description: "data to create the token or a new user"
        required: true
        schema:
          $ref: "#/definitions/Body"
      responses:
        200:
          description: "Successfully returned a token"
          schema:
            type: "object"
            required:
            - "token"
            properties:
              error:
                type: "boolean"
              token:
                type: "string"
              briefFilled:
                type: "boolean"
            example:
              error: false
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVjOWVmZGRlMGJiOWRlZmM5YzIxOGM2NSIsImlzQWRtaW4iOmZhbHNlLCJ2ZXJpZmllZCI6dHJ1ZSwiaWF0IjoxNTUzOTM0NTA2LCJleHAiOjE1NTM5NzA1MDZ9.ron2Gh--Jym8p_8Iqu4gbuo6rO1LE8DKzAdrezwIWKY"
        400:
          description: "Error in data of Body. Invalid Request. No token in body"
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "No token in body!"
        401:
          description: "Error in token. Un-authorized! Expired token, Issuer Incorrect\
            \ or Incorrect Client."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Expired Token!"
      x-swagger-router-controller: "Default"
    put:
      description: "Edits the Profile of the Logged In-User."
      operationId: "usersPUT"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "JWT token set in header"
        required: true
        type: "string"
      - in: "body"
        name: "User"
        description: "The user Profile to Edit."
        required: false
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "User Profile Edited."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              user:
                type: "object"
                properties: {}
            example:
              error: false
              user:
                _id: "5c9efdde0bb9defc9c218c65"
                firstName: "Udit"
                lastName: "Jain"
                email: "jainudit.012@gmail.com"
                emailVerified: true
                profilePicture: "https://media.licdn.com/dms/image/C5103AQG_1EUajUCTXg/profile-displayphoto-shrink_100_100/0?e=1559174400&v=beta&t=WQDJMxQCMYLl4rJ2yYszirYi-rVnd2q1xvtD5greIyY"
                fullName: "Udit Jain"
                sub: "linkedin|oP4sPrYxft"
                purpose: "my purpose"
                brief: "thi is my brief"
                isAdmin: true
                isFormFilled: true
                isActive: true
                createdAt: "2019-03-30T05:25:50.413Z"
                updatedAt: "2019-03-30T06:25:13.266Z"
        401:
          description: "Not logged in or no token send in headers."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Not Logged In!"
        404:
          description: "User not found in database."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Could not find the User!"
        500:
          description: "User Profile Not Updated"
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Could not Update the Profile!"
      x-swagger-router-controller: "Default"
  /users/me:
    get:
      description: "Returns Profile of the Logged in user."
      operationId: "usersMeGET"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "JWT token set in header"
        required: true
        type: "string"
      responses:
        200:
          description: "User Profile Found."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              user:
                type: "object"
                properties: {}
              projects:
                type: "number"
              opportunities:
                type: "number"
              funds:
                type: "number"
            example:
              error: false
              user:
                _id: "5c9efdde0bb9defc9c218c65"
                firstName: "Udit"
                lastName: "Jain"
                email: "jainudit.012@gmail.com"
                emailVerified: true
                profilePicture: "https://media.licdn.com/dms/image/C5103AQG_1EUajUCTXg/profile-displayphoto-shrink_100_100/0?e=1559174400&v=beta&t=WQDJMxQCMYLl4rJ2yYszirYi-rVnd2q1xvtD5greIyY"
                fullName: "Udit Jain"
                sub: "linkedin|oP4sPrYxft"
                purpose: "my purpose"
                brief: "my brief"
                isAdmin: true
                isFormFilled: true
                isActive: true
                createdAt: "2019-03-30T05:25:50.413Z"
                updatedAt: "2019-03-30T06:25:13.266Z"
              projects: 2
              opportunities: 0
              funds: 0
        401:
          description: "Not logged in or no token send in headers."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Not Logged In!"
        404:
          description: "User not found in database."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Could not find the User!"
      x-swagger-router-controller: "Default"
  /users/{userId}:
    get:
      description: "Returns Profile of the specified User for the Admin."
      operationId: "usersUserIdGET"
      parameters:
      - name: "userId"
        in: "path"
        description: "Id of the User that is to searched."
        required: true
        type: "string"
      - name: "Authorization"
        in: "header"
        description: "JWT token set in header"
        required: true
        type: "string"
      responses:
        200:
          description: "User Profile Found."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              user:
                type: "object"
                properties: {}
            example:
              error: false
              user:
                _id: "5c9efdde0bb9defc9c218c65"
                firstName: "Udit"
                lastName: "Jain"
                email: "jainudit.012@gmail.com"
                emailVerified: true
                profilePicture: "https://media.licdn.com/dms/image/C5103AQG_1EUajUCTXg/profile-displayphoto-shrink_100_100/0?e=1559174400&v=beta&t=WQDJMxQCMYLl4rJ2yYszirYi-rVnd2q1xvtD5greIyY"
                fullName: "Udit Jain"
                sub: "linkedin|oP4sPrYxft"
                purpose: "my purpose"
                brief: "my brief"
                isAdmin: true
                isFormFilled: true
                isActive: true
                createdAt: "2019-03-30T05:25:50.413Z"
                updatedAt: "2019-03-30T06:25:13.266Z"
        403:
          description: "Not Admin User."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Not Authorized!"
        404:
          description: "User not found in database."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Could not find the User!"
      x-swagger-router-controller: "Default"
  /projects:
    get:
      description: "Should get All Project if User is Admin."
      operationId: "projectsGET"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "JWT token set in header"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "Limits number of documents returned"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "No. of documents to skip"
        required: false
        type: "string"
      - name: "sector"
        in: "query"
        required: false
        type: "string"
        enum:
        - "HEALTHCARE_FITNESS"
        - "FIN_TECH"
        - "E_COMMERCE"
        - "CONSUMER"
        - "FOOD"
        - "EDUCATION"
        - "HR"
      - name: "projectStage"
        in: "query"
        required: false
        type: "string"
        enum:
        - "IDEA"
        - "MVP"
        - "GROWTH"
        - "BETA"
        - "GROWING"
      responses:
        200:
          description: "Returns list of All the Projects"
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              projects:
                type: "array"
                items:
                  type: "object"
                  properties: {}
            example:
              error: false
              projects:
              - _id: "5c9efdde0bb9defc9c218c65"
                firstName: "Udit"
                lastName: "Jain"
                email: "jainudit.012@gmail.com"
                emailVerified: true
                profilePicture: "https://media.licdn.com/dms/image/C5103AQG_1EUajUCTXg/profile-displayphoto-shrink_100_100/0?e=1559174400&v=beta&t=WQDJMxQCMYLl4rJ2yYszirYi-rVnd2q1xvtD5greIyY"
                fullName: "Udit Jain"
                sub: "linkedin|oP4sPrYxft"
                purpose: "my purpose"
                brief: "thi is my brief"
                isAdmin: true
                isFormFilled: true
                isActive: true
                createdAt: "2019-03-30T05:25:50.413Z"
                updatedAt: "2019-03-30T06:25:13.266Z"
        403:
          description: "Not an Admin User."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "You do not have Correct Permissions!"
      x-swagger-router-controller: "Default"
    post:
      description: "Creates Project if User is Logged In."
      operationId: "projectsPOST"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "JWT token set in header"
        required: true
        type: "string"
      - in: "body"
        name: "Project"
        description: "The Project Data from the User in request Body."
        required: true
        schema:
          $ref: "#/definitions/Project"
      responses:
        200:
          description: "Creates the Project."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              project:
                type: "object"
                properties: {}
              msg:
                type: "string"
            example:
              error: false
              project:
                projectName: "Start Universe"
                projectNature: "SERVICE"
                sector: "EDUCATION"
                startupStage: "IDEA"
                teamSize: "10"
                acceptingFunds: true
                pitch: "my pitch for this project that will make it a success."
                problemSolvingDescription: "my description"
                createdBy:
                  _id: "5c9efdde0bb9defc9c218c65"
                  firstName: "Udit"
                  lastName: "Jain"
                  email: "jainudit.012@gmail.com"
                  emailVerified: true
                  profilePicture: "https://media.licdn.com/dms/image/C5103AQG_1EUajUCTXg/profile-displayphoto-shrink_100_100/0?e=1559174400&v=beta&t=WQDJMxQCMYLl4rJ2yYszirYi-rVnd2q1xvtD5greIyY"
                  fullName: "Udit Jain"
                  sub: "linkedin|oP4sPrYxft"
                  purpose: "my purpose is this?"
                  brief: "my brief"
                  isAdmin: true
                  isFormFilled: true
                  isActive: true
                  createdAt: "2019-03-30T05:25:50.413Z"
                  updatedAt: "2019-04-01T09:38:11.109Z"
                isApproved: false
                createdByName: "Udit Jain"
                projectApprovalStatus: "ADDED"
                createdAt: "2019-04-03T10:00:02.116Z"
                updatedAt: "2019-04-03T10:00:02.116Z"
                _id: "5ca48422b7b102e96f3fae9b"
              msg: "New Project Successfully Created."
        400:
          description: "Bad request. Missing Some Fields."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "No Project Stage provided!"
        401:
          description: "If user is not Logged in."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Cannot create Project. Must be logged in!"
        404:
          description: "No User found In Database"
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Could not find the User!"
      x-swagger-router-controller: "Default"
...
  /projects/approve/{projectId}:
    put:
      description: "Approves Project if User is Admin."
      operationId: "projectsApproveProjectIdPUT"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "JWT token set in header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        description: "Id of the Project that is to be changed"
        required: true
        type: "string"
      - in: "body"
        name: "Approval Status"
        description: "Status that is to be set for the Project."
        required: false
        schema:
          $ref: "#/definitions/Approval Status"
      responses:
        200:
          description: "Updates the Approval Status of the Project."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              project:
                type: "object"
                properties: {}
              msg:
                type: "string"
            example:
              error: false
              project:
                _id: "5ca48422b7b102e96f3fae9b"
                projectName: "Universe's Model"
                projectNature: "SERVICE"
                sector: "EDUCATION"
                startupStage: "IDEA"
                teamSize: "10"
                pitch: "my pitch for this project that will make it a success."
                acceptingFunds: true
                isApproved: "false"
                projectApprovalStatus: "IN_REVIEW"
                problemSolvingDescription: "my description"
                createdByName: "Udit Jain"
                createdAt: "2019-04-03T10:00:02.116Z"
                updatedAt: "2019-04-03T10:39:45.165Z"
              msg: "Successfully Updated Approval Status."
        400:
          description: "Bad request. Incorrect Data Value."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Status can be only IN_REVIEW, APPROVED or REJECTED"
        403:
          description: "If user is not Admin."
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Not An Admin!"
        404:
          description: "No User or No Searched Project found In Database"
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Could Not find the Project!"
        500:
          description: "Document Updation Error"
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              reason:
                type: "string"
            example:
              error: true
              reason: "Could Not Update the Project!"
      x-swagger-router-controller: "Default"
...
  /ideas:
    get:
      description: "Gets the Latest 3 Approved Projects"
      operationId: "ideasGET"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "JWT token set in header"
        required: false
        type: "string"
      responses:
        200:
          description: "Return List of 3 Latest Approved Projects"
          schema:
            type: "object"
            properties:
              error:
                type: "boolean"
              trending:
                type: "array"
                items:
                  type: "object"
                  properties: {}
            example:
              error: false
              trending: []
      x-swagger-router-controller: "Default"
...
definitions:
...
  User:
    type: "object"
    properties:
      purpose:
        type: "string"
      brief:
        type: "string"
      userType:
        type: "string"
        enum:
        - "Investor"
        - "Mentor"
        - "Consulting_Firm"
        - "IT_Services"
        - "Investment_Advisor"
        - "Startup_Founder"
      image:
        type: "string"
    example:
      purpose: "my purpose"
      brief: "this is my brief"
...
  Body:
    type: "object"
    required:
    - "purpose"
    - "token"
    properties:
      token:
        type: "string"
        description: "token from `OAuth`"
      purpose:
        type: "string"
        description: "Should be`Sent one time only,  Can be FOUNDER, INVESTOR or MENTOR`."
      brief:
        type: "string"
...